🐛fix: evitar eliminar vector_store inexistente
Se agregó una validación para comprobar si `vector_store_id` existe antes de realizar operaciones asociadas. Esto previene posibles errores al intentar acceder a propiedades nulas o inexistentes.

+++ app/Http/Controllers/SubjectController.php
@@ -103,8 +103,10 @@ class SubjectController extends Controller
         $subject = Subject::find($id);
 
         $assistantService->deleteAssistant($subject->assistant_id);
-        $filesService->deleteFiles($subject->vector_store_id);
-        $filesService->deleteVectorStore($subject->vector_store_id);
+        if ($subject->vector_store_id) {
+            $filesService->deleteFiles($subject->vector_store_id);
+            $filesService->deleteVectorStore($subject->vector_store_id);
+        }
         $subject->delete();
 
         return redirect()->route('subjects.index');
